# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QuoteMode = typeof QuoteMode[keyof typeof QuoteMode];


export const QuoteMode = {
  residential: 'residential',
  commercial: 'commercial',
} as const;

export interface Quote {
  id: string;
  name: string;
  contact: string;
  message: string;
  mode: QuoteMode;
  createdAt: string;
}

export type QuoteInputMode = typeof QuoteInputMode[keyof typeof QuoteInputMode];


export const QuoteInputMode = {
  residential: 'residential',
  commercial: 'commercial',
} as const;

export interface QuoteInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 5
     * @maxLength 160
     */
  contact: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
  mode: QuoteInputMode;
}

export interface ErrorResponse {
  error: string;
}