# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteMode } from './quoteMode';

export interface Quote {
  id: string;
  name: string;
  contact: string;
  message: string;
  mode: QuoteMode;
  createdAt: Date;
}