# lib/api-zod/src/generated/types/quoteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteInputMode } from './quoteInputMode';

export interface QuoteInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 5
     * @maxLength 160
     */
  contact: string;
  /**
     * @minLength 10
     * @maxLength 2000
     */
  message: string;
  mode: QuoteInputMode;
}